Move serializer from hardware app to employee app
